requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
grpcio==1.64.1
huggingface-hub==0.23.2
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
-e git+https://github.com/chongdashu/stable-diffusion-imagic@e904e691a8490b35de73ef52e3e08b2495442055#egg=latent_diffusion
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.1
pyDeprecate==0.3.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.5
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sympy==1.12.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tokenizers==0.19.1
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
